Adds argument to select version to compute
Changes the flag to compute the correlation-based version of the parameters by an argument that lets choose among the three versions. This is a more consistent behavior than using an identity matrix for computing the agnostic version.
Then, if one wants to compute it, the covariance matrix can be omitted.