Corrects statement about axes interpretation
The interpretation of the axes in the latent space is not that their angles represent the correlations (that happens in the test space). Therefore, that sentence in the introduction is changed to a more neutral one, that introduces this topic, which will be clarified later on in the results and discussion.